requirements.txt
PyMuPDF
sentence-transformers
faiss-cpu
transformers
torch
fastapi
uvicorn
pydantic
scikit-learn
python-multipart
gunicorn
